worker/requirements.txt
# Async task queue
celery[redis]==5.4.0
redis==5.1.1

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Neo4j
neo4j==5.24.0

# AWS SDK
boto3==1.35.0
botocore==1.35.0

# Graph processing
networkx==3.3

# HTTP
httpx==0.27.2
requests==2.32.3

# Data
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Logging
structlog==24.4.0
tenacity==9.0.0

# AI providers (Day 4)
google-genai>=0.8.0
anthropic==0.34.0
openai==1.45.0

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
